feat: first structure tensor version
