+++ CMakePresets.json
@@ -14,6 +14,10 @@
                 "value": "x64",
                 "strategy": "external"
             },
+            "toolchainFile": "$env{VCPKG_ROOT}/scripts/buildsystems/vcpkg.cmake",
+            "cacheVariables": {
+                "VCPKG_TARGET_TRIPLET": "x64-windows-static"
+            },
             "condition": {
                 "type": "equals",
                 "lhs": "${hostSystemName}",
